third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-data-in-workers-expected.txt

[Worker] Test createImageBitmap with data in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ./resources/canvas-createImageBitmap-data-in-workers.js
PASS [Worker] Promise fulfilled.
PASS [Worker] bitmapWidth is 50
PASS [Worker] bitmapHeight is 50
PASS successfullyParsed is true

TEST COMPLETE
